+++ query-planner-js/src/composedSchema/index.ts
@@ -1,2 +1,3 @@
 export { buildComposedSchema } from './buildComposedSchema';
+export { toAPISchema } from './toAPISchema';
 export * from './metadata';